authorization done

+++ routes/web.php
@@ -1,36 +1,38 @@
 <?php
 
 use App\Http\Controllers\JobController;
+use App\Http\Controllers\RegisteredUserController;
+use App\Http\Controllers\SessionController;
+use Illuminate\Contracts\Session\Session;
 use Illuminate\Support\Facades\Route;
 
 // Home page 
 Route::view('/','home');
 
+
 Route ::controller(JobController::class)->group(function(){
-
-    // Index (List all jobs)
 Route::get('/jobs','index');
-
-// Create job (Show the form to create a job)
 Route::get('/jobs/create','create');
-
-// Show job details by ID
 Route::get('/jobs/{job}','show');  
-
-// Store a new job
-Route::post('/jobs','store');
-
-// Edit job (Show the form to edit a job)
-Route::get('/jobs/{job}/edit','edit');  
-
-// Update a job
-Route::patch('/jobs/{job}','update');  
-
-// Delete a job
-Route::delete('/jobs/{job}','destroy');  
-
+Route::post('/jobs','store')->middleware('auth');
+Route::get('/jobs/{job}/edit','edit')->middleware('auth')->can('can:edit-job,job' );
+Route::patch('/jobs/{job}','update')->middleware('auth')->can('can:edit-job,job' );
+Route::delete('/jobs/{job}','destroy')->middleware('auth')->can('can:edit-job,job' );  
 });
+
 
 
 // Contact page
 Route::view('/contact','contact');
+
+
+//Auth
+
+
+Route::get('/register',[RegisteredUserController::class,'create']);
+Route::post('/register',[RegisteredUserController::class,'store']);
+
+Route::get('/login',[SessionController::class,'create']);
+Route::post('/login',[SessionController::class,'store']);
+Route::post('/logout',[SessionController::class,'destroy']);
+
